finished htmx flask mini project

+++ python-web-dev/flask-gunpla-monolith-htmx/app/routes/main.py
@@ -1,4 +1,5 @@
-from flask import Blueprint, render_template, redirect, url_for, flash, request
+# app/routes/main.py
+from flask import Blueprint, render_template, redirect, url_for, flash, request, abort
 from app.models.gunpla import Gunpla
 from app.forms.gunpla import GunplaForm
 from app import db
@@ -7,7 +8,7 @@ bp = Blueprint('main', __name__)
 
 @bp.route('/')
 def index():
-    gunplas = Gunpla.query.all()
+    gunplas = db.session.execute(db.select(Gunpla)).scalars().all()
     return render_template('gunpla/index.html', gunplas=gunplas)
 
 @bp.route('/create', methods=['GET', 'POST'])
@@ -28,7 +29,9 @@ def create():
 
 @bp.route('/edit/<int:gunpla_id>', methods=['GET', 'POST'])
 def edit(gunpla_id):
-    gunpla = Gunpla.query.get_or_404(gunpla_id)
+    gunpla = db.session.get(Gunpla, gunpla_id)
+    if gunpla is None:
+        abort(404)
     form = GunplaForm(obj=gunpla)
     if form.validate_on_submit():
         gunpla.name = form.name.data
@@ -42,7 +45,9 @@ def edit(gunpla_id):
 
 @bp.route('/delete/<int:gunpla_id>', methods=['POST'])
 def delete(gunpla_id):
-    gunpla = Gunpla.query.get_or_404(gunpla_id)
+    gunpla = db.session.get(Gunpla, gunpla_id)
+    if gunpla is None:
+        abort(404)
     db.session.delete(gunpla)
     db.session.commit()
     flash('Gunpla model deleted successfully!', 'success')
@@ -50,11 +55,15 @@ def delete(gunpla_id):
 
 @bp.route('/gunpla/<int:gunpla_id>/edit-form')
 def get_edit_form(gunpla_id):
-    gunpla = Gunpla.query.get_or_404(gunpla_id)
+    gunpla = db.session.get(Gunpla, gunpla_id)
+    if gunpla is None:
+        abort(404)
     form = GunplaForm(obj=gunpla)
     return render_template('gunpla/edit_form.html', form=form, gunpla=gunpla)
 
 @bp.route('/gunpla/<int:gunpla_id>')
 def get_gunpla_row(gunpla_id):
-    gunpla = Gunpla.query.get_or_404(gunpla_id)
+    gunpla = db.session.get(Gunpla, gunpla_id)
+    if gunpla is None:
+        abort(404)
     return render_template('gunpla/gunpla_row.html', gunpla=gunpla)
